Decouple child-exited from eof
Currently vte delays the "child-exited" signal to until after eos has
been received. This is wrong and caused a regression.

vte shouldn't tell client terminals when to exit, if the children has
exited, vte should send the "child-exited" signal. Period.

If the client decides to wait for "eof", that's up to the client
terminal.

On Unix if the child spawns grandchildren, their output will be attached
to the controlling terminal (unless a tool like nohup is used),
therefore the eos won't happen until all the grandchildren have closed
their output.

For example this command will keep producing output after the shell has
exited:

  sh -c 'for i in $(seq 1 10); do echo $i; sleep 1; done' & exit

vte assumes an eos will always occur, which is why if a shell spawns any
grandchildren, exiting the shell doesn't close the terminal: vte would
wait for all the grandchildren to exit, which from the point of view of
the user it seems the terminal is hanging forever.

To workaround this issue, a 5-second timer was introduced in 058adf5f
(pty: Fix indefinite wait for EOS after child-exited, 2019-12-01), which
was later reduced to 2 seconds, but this is just a hack.

vte should not be waiting for an eos before sending "child-exited".

In the past vte used to send the "child-exited" immediately, which is
the correct desired behavior, but that changed in commit 7888602c (lib:
Rework child exit and EOF handling, 2019-11-17).

The only correct change in that commit is removing unset_pty() from
child_watch_done() which does allow the code to process remaining data
in the buffers, even from grandchildren that take several seconds to
finish.

Delaying "child-exited" was not needed to fix that problem and
introduced the regression that can be easily triggered by simply doing:

  sleep infinity & exit

Decoupling "child-exited" from "eof" gets rid of a lot of unnecessary
complexity, and also eliminates the need for the hacky timer.

With this change even the output from grandchildren processes that last
several seconds is processed correctly:

  xfce4-terminal --disable-server --hold \
  --command="bash -i -c \"sh -c 'for i in \$(seq 1 10); do echo \$i; sleep 1; done' & exit\""

And of course, the vastly most common case of terminals running shells
that want to exit immediately works correctly.

Fixes #204.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>