vte: send child-exited when child exits
The commit 7888602c (lib: Rework child exit and EOF handling,
2019-11-17) introduced a regression when grandchild processes are
present.

The bug is easy to reproduce:

  sleep 10 & exit

The reason is that the `child-exited` signal is delayed to until after
eos has been received, but that doesn't happen until after all the
grandchild processes have exited.

The `child-exited` signal should be emitted when the child has exited.
Period.

If we do that all the complexity of delaying the signal, including the
hacks, can be removed.

Fixes #204.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>