pty: Fix exit delay in presence of a grandchild process
When we notice that the child has quit, keep processing incoming data only as
long as there's data immediately available, and only up to a certain amount
which is presumably larger than the kernel's tty buffer size.

This should be a more robust solution than the timeout-based one it replaces,
in order to consume all the data produced by the immediate child and then to
emit the child-exited signal without further delay.

Fixes: https://gitlab.gnome.org/GNOME/vte/-/issues/2627