vte: cleanup unused stuff
In commit 1bcf4bca (pty: Fix exit delay in presence of a grandchild
process, 2023-08-04) the call to Terminal::queue_child_exited() was
removed, but not the method.

If we remove that method, a bunch of other methods can be removed as
well, in addition to m_child_exit_status.

Plus, in commit 7888602c (lib: Rework child exit and EOF handling,
2019-11-17) the definition of the method Terminal::pty_channel_eof() was
removed, but not the declaration.

Let's remove all this unused stuff.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>