vte: don't read extra data unnecessarily
Commit 1bcf4bca (pty: Fix exit delay in presence of a grandchild
process, 2023-08-04) introduced a bunch of code that isn't necessary.

If the terminal is interested in the data that is incoming right when
the child is exiting, then it can simply ignore `child-exited` and the
data will be processed normally.

There's no need to read a chunk of 64 k data before sending
`child-exited` because the data will be read afterwards.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>